requirements.txt
pydantic~=2.9
puzpy~=0.2.6
lxml~=5.3
unidecode~=1.3
pytest~=8.3
openai~=1.55
python-dotenv~=1.0
